pdf_chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gradio==4.7.1
ollama==0.1.7
pypdf2==3.0.1
sentence-transformers==2.2.2
faiss-cpu>=1.12.0
numpy>=1.24.3
python-multipart==0.0.6
pydantic==2.5.0

